server: add current to obtain executor handle
It aims to reduce argument passing complex.
